feat: simplify dom package
